requirements.txt
blinker==1.8.2 ; python_version >= "3.10" and python_version < "4"
cachecontrol==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.17.0 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4"
cloudevents==1.9.0 ; python_version >= "3.10" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4" and platform_system == "Windows"
cryptography==43.0.1 ; python_version >= "3.10" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.10" and python_version < "4"
firebase-admin==6.5.0 ; python_version >= "3.10" and python_version < "4.0"
firebase-functions==0.4.2 ; python_version >= "3.10" and python_version < "4.0"
flask-cors==5.0.0 ; python_version >= "3.10" and python_version < "4.0"
flask==3.0.3 ; python_version >= "3.10" and python_version < "4"
functions-framework==3.8.1 ; python_version >= "3.10" and python_version < "4"
google-api-core==2.19.2 ; python_version >= "3.10" and python_version < "4.0"
google-api-core[grpc]==2.19.2 ; python_version >= "3.10" and python_version < "4.0"
google-api-python-client==2.143.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.34.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-firestore==2.18.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-storage==2.18.2 ; python_version >= "3.10" and python_version < "4.0"
google-crc32c==1.6.0 ; python_version >= "3.10" and python_version < "4.0"
google-events==0.12.0 ; python_version >= "3.10" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.10" and python_version < "4.0"
grpcio-status==1.62.3 ; python_version >= "3.10" and python_version < "4.0"
grpcio==1.66.1 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.10" and python_version < "4" and platform_system != "Windows"
httplib2==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.8 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.10" and python_version < "4"
jinja2==3.1.4 ; python_version >= "3.10" and python_version < "4"
markupsafe==2.1.5 ; python_version >= "3.10" and python_version < "4"
mock-firestore==0.11.0 ; python_version >= "3.10" and python_version < "4.0"
msgpack==1.0.8 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.10" and python_version < "4"
proto-plus==1.24.0 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.25.4 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
pyjwt[crypto]==2.9.0 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.1.4 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.10" and python_version < "4.0"
watchdog==5.0.2 ; python_version >= "3.10" and python_version < "4"
werkzeug==3.0.4 ; python_version >= "3.10" and python_version < "4"
